Improvement: Better display of shoppingcart bookings on mybookings.php

+++ mybookings.php
@@ -44,10 +44,10 @@ echo $OUTPUT->header();
 
 echo $OUTPUT->heading(get_string('mybookingoptions', 'mod_booking'));
 
-if (class_exists('local_shopping_cart\shopping_cart')) {
+echo shortcodes::mycourselist('', [], '', (object)[], fn($a) => $a);
+
+if (class_exists('local_shopping_cart\shopping_cart') && get_config('booking', 'displayshoppingcarthistory')) {
     echo local_shopping_cart\shortcodes::shoppingcarthistory('', [], '', (object)[], fn($a) => $a);
 }
-
-echo shortcodes::mycourselist('', [], '', (object)[], fn($a) => $a);
 
 echo $OUTPUT->footer();
